// src/app/interfaces/projet.ts
export interface Projet {
    id?:number;
    icon?: string;
    outils?: string;
    nomIcon?: string;
    imageJombo?: string;
    imageUrl?: string;
    imageSeo?: string;
    titre?: string;
    resume?: string;
    premier?: string;
    deux?: string;
    trois?: string;
    quatre?: string;
    cinq?: string;
    six?: string;
    sept?: string;
    huit?: string;
    neuf?: string;
    dix?: string;
}